refactor: fully separate OS kernel and componentize logic
This commit completes the deep architectural refactoring.

- All core OS shell files (`App.tsx`, `types.ts`, `constants.tsx`,
  core UI components, contexts, and themes) are now encapsulated
  within the new `window/` (kernel) directory. The root directory
  is now clean, containing only project config and the main entry
  point.

- All import paths throughout the entire application have been
  updated to reflect this new structure.

- The `useWindowManager` hook now correctly handles dependencies
  and logic for the windowing system.

- Redundant legacy files like `components/constants.tsx` have been
  removed.
