+++ components/apps/AppStoreApp.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { AppDefinition, AppComponentProps } from '../../types';
-import { StoreIcon, RefreshIcon } from '../../constants';
-import * as FsService from '../../services/filesystemService';
-import { APP_DEFINITIONS } from '.';
+import { AppDefinition, AppComponentProps } from '../../../window/types';
+import { AppStoreIcon, RefreshIcon } from '../../../window/constants';
+import * as FsService from '../../../services/filesystemService';
+import { APP_DEFINITIONS } from '..';
 
 const AppStoreApp: React.FC<AppComponentProps> = ({ setTitle, initialData }) => {
     const [installedAppIds, setInstalledAppIds] = useState<Set<string>>(new Set());
@@ -53,7 +53,7 @@ const AppStoreApp: React.FC<AppComponentProps> = ({ setTitle, initialData }) => 
         <div className="p-6 text-zinc-200 h-full overflow-y-auto custom-scrollbar bg-zinc-900">
             <div className="flex items-center justify-between mb-6">
                  <div className="flex items-center">
-                    <StoreIcon className="w-10 h-10 text-blue-400 mr-4" />
+                    <AppStoreIcon className="w-10 h-10 text-blue-400 mr-4" />
                     <div>
                         <h1 className="text-2xl font-semibold text-white">App Store</h1>
                         <p className="text-sm text-zinc-400">Discover and install new applications.</p>
@@ -105,7 +105,7 @@ const AppStoreApp: React.FC<AppComponentProps> = ({ setTitle, initialData }) => 
 export const appDefinition: AppDefinition = {
     id: 'appStore',
     name: 'App Store',
-    icon: StoreIcon,
+    icon: AppStoreIcon,
     component: AppStoreApp,
     defaultSize: { width: 750, height: 550 },
     isPinnedToTaskbar: true,
