+++ index.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './window/App';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
