+++ window/theme/default.ts
@@ -1,5 +1,5 @@
 
-import { Theme } from '../../types';
+import { Theme } from '../types';
 
 export const defaultTheme: Theme = {
   id: 'default',
