+++ window/theme/light.ts
@@ -1,5 +1,5 @@
 
-import { Theme } from '../../types';
+import { Theme } from '../types';
 
 export const lightTheme: Theme = {
   id: 'light',
