+++ window/types.ts
